refactor: move lifecycle playback scenarios
